Add location transfer with picture data
To implement this feature, Fine location permission is added in the manifest file, location in latitude and longitude with 4
decimal precision is added as the last 8 + 9 bytes in the unencrypted byte array, location boolean is added to syncdata.
Also, the latest SegmentFetcher from jNDN is placed with little modification and segment verification is bypassed.